refactor: Rearrange the platform creation and further content
Mainly to avoid top-level definitions, that will make it simpler preventing dependency loops (and already helped me detecting some accidental imports)
