+++ src/qibolab/backends.py
@@ -7,11 +7,32 @@ from qibo.config import raise_error
 from qibo.models import Circuit
 from qibo.result import MeasurementOutcomes
 
-from qibolab import ExecutionParameters
-from qibolab import __version__ as qibolab_version
-from qibolab import create_platform
 from qibolab.compilers import Compiler
-from qibolab.platform import Platform
+from qibolab.execution_parameters import ExecutionParameters
+from qibolab.platform import Platform, create_platform
+from qibolab.platform.load import available_platforms
+from qibolab.version import __version__ as qibolab_version
+
+
+def execute_qasm(circuit: str, platform, initial_state=None, nshots=1000):
+    """Executes a QASM circuit.
+
+    Args:
+        circuit (str): the QASM circuit.
+        platform (str): the platform where to execute the circuit.
+        initial_state (:class:`qibo.models.circuit.Circuit`): Circuit to prepare the initial state.
+                If ``None`` the default ``|00...0>`` state is used.
+        nshots (int): Number of shots to sample from the experiment.
+
+    Returns:
+        ``MeasurementOutcomes`` object containing the results acquired from the execution.
+    """
+    from qibolab.backends import QibolabBackend
+
+    circuit = Circuit.from_qasm(circuit)
+    return QibolabBackend(platform).execute_circuit(
+        circuit, initial_state=initial_state, nshots=nshots
+    )
 
 
 class QibolabBackend(NumpyBackend):
@@ -146,3 +167,24 @@ class QibolabBackend(NumpyBackend):
                 gate.result.backend = self
                 gate.result.register_samples(np.array(samples).T)
         return results
+
+
+class MetaBackend:
+    """Meta-backend class which takes care of loading the qibolab backend."""
+
+    @staticmethod
+    def load(platform: str):
+        """Loads the backend.
+
+        Args:
+            platform (str): Name of the platform to load.
+        Returns:
+            qibo.backends.abstract.Backend: The loaded backend.
+        """
+        from qibolab.backends import QibolabBackend
+
+        return QibolabBackend(platform=platform)
+
+    def list_available(self) -> dict:
+        """Lists all the available qibolab platforms."""
+        return {platform: True for platform in available_platforms()}
