+++ src/qibolab/version.py
@@ -1,0 +1,3 @@
+import importlib.metadata as im
+
+__version__ = im.version(__package__)
